+++ Lib/MoneytreeLinkCoreKit.xcframework/ios-arm64/MoneytreeLinkCoreKit.framework/Headers/MTLApplicationDelegate.h
@@ -68,6 +68,37 @@ typedef void(^MTRestoreUserActivityCompletionBlock)(NSArray<id<UIUserActivityRes
         presentFrom:(UIViewController *)viewController
          completion:(MTLinkCompletionBlock)completion;
 
+
+/// Expected to be called by the SDK's client to let the framework handle URLs
+/// when using the scene-based lifecycle.
+/// Call this from your `UIWindowSceneDelegate`'s `scene(_:openURLContexts:)` method.
+///
+/// @param scene The scene that received the URL.
+/// @param URLContexts A set of `UIOpenURLContext` objects. The SDK will inspect the first URL.
+/// @return `YES` if the SDK is expecting the URL and handling it.
+- (BOOL)scene:(UIScene *)scene
+openURLContexts:(NSSet<UIOpenURLContext *> *)URLContexts API_AVAILABLE(ios(13.0));
+
+/// Expected to be called by the SDK's client to let the framework check whether
+/// the URL is expected for universal links when using the scene-based lifecycle.
+/// Call this from your `UIWindowSceneDelegate`'s `scene(_:continue:)` method.
+///
+/// @param scene The scene continuing the activity.
+/// @param userActivity The activity object containing the data associated with the task the user was performing.
+/// @param viewController A view controller where a webview completing the Login Link process will be shown.
+/// @param completion A block that is called after the userActivity is handled.
+///                   If the universal link can not be handled, completion will not be called.
+///
+///                   The completion contains a null error if the activity is successfully handled.
+///                   The completion contains a non-null error if the activity failed to be handled.
+/// @return `YES` if the SDK is expecting the URL and handling it.
+- (BOOL)scene:(UIScene *)scene
+continueUserActivity:(NSUserActivity *)userActivity
+         presentFrom:(UIViewController *)viewController
+   completion:(MTLinkCompletionBlock)completion API_AVAILABLE(ios(13.0))
+NS_SWIFT_NAME(scene(_:continue:presentFrom:completion:));
+
+
 @end
 
 NS_ASSUME_NONNULL_END